refactor: fix package hierarchy
