+++ src/main/java/edu/kit/kastel/sdq/lissa/ratlr/promptoptimizer/AutomaticPromptOptimizer.java
@@ -24,8 +24,8 @@ import edu.kit.kastel.sdq.lissa.ratlr.classifier.ClassificationTask;
 import edu.kit.kastel.sdq.lissa.ratlr.configuration.ModuleConfiguration;
 import edu.kit.kastel.sdq.lissa.ratlr.elementstore.SourceElementStore;
 import edu.kit.kastel.sdq.lissa.ratlr.elementstore.TargetElementStore;
-import edu.kit.kastel.sdq.lissa.ratlr.evaluator.BruteForceEvaluator;
-import edu.kit.kastel.sdq.lissa.ratlr.evaluator.Evaluator;
+import edu.kit.kastel.sdq.lissa.ratlr.promptoptimizer.evaluator.BruteForceEvaluator;
+import edu.kit.kastel.sdq.lissa.ratlr.promptoptimizer.evaluator.Evaluator;
 import edu.kit.kastel.sdq.lissa.ratlr.knowledge.TraceLink;
 import edu.kit.kastel.sdq.lissa.ratlr.promptmetric.Metric;
 import edu.kit.kastel.sdq.lissa.ratlr.promptoptimizer.samplestrategy.FirstSampler;
